Fixed warnings + RANSAC confidence to double

+++ modules/calib3d/include/opencv2/calib3d.hpp
@@ -159,7 +159,7 @@ CV_EXPORTS_W bool solvePnPRansac( InputArray objectPoints, InputArray imagePoint
                                   InputArray cameraMatrix, InputArray distCoeffs,
                                   OutputArray rvec, OutputArray tvec,
                                   bool useExtrinsicGuess = false, int iterationsCount = 100,
-                                  float reprojectionError = 8.0, float confidence = 0.99,
+                                  float reprojectionError = 8.0, double confidence = 0.99,
                                   OutputArray inliers = noArray(), int flags = ITERATIVE );
 
 //! initializes camera matrix from a few 3D points and the corresponding projections.
